@@ -1,37 +1,37 @@
 module main
 
-go 1.21
+go 1.22.0
+
+toolchain go1.22.2
 
 require (
-	github.com/aws/aws-sdk-go v1.45.3
-	github.com/google/go-tpm v0.9.0
+	github.com/aws/aws-sdk-go-v2 v1.27.0
+	github.com/aws/aws-sdk-go-v2/config v1.27.16
+	github.com/aws/aws-sdk-go-v2/service/sts v1.28.10
+	github.com/foxboron/go-tpm-keyfiles v0.0.0-20240525122353-0883da4eb332
+	github.com/google/go-tpm v0.9.1-0.20240514145214-58e3e47cd434
 	github.com/google/go-tpm-tools v0.4.0
-	github.com/gorilla/schema v1.2.0
-	github.com/salrashid123/aws_hmac/stsschema v0.0.0-20230907023647-8ea8949a03f4
-	github.com/salrashid123/aws_hmac/tpm v0.0.0-20230907030222-03be18507813
-	github.com/salrashid123/aws_hmac/tpm/signer v0.0.0-20230907024207-0899a3ac266a
-	github.com/salrashid123/aws_hmac/tpm/signer/v4 v4.0.0-20230907024207-0899a3ac266a
-// github.com/salrashid123/aws_hmac/stsschema v0.0.0
-// github.com/salrashid123/aws_hmac/tpm v0.0.0
-// github.com/salrashid123/aws_hmac/tpm/signer v0.0.0
-// github.com/salrashid123/aws_hmac/tpm/signer/v4 v4.0.0
+	github.com/gorilla/schema v1.3.0
+	github.com/salrashid123/aws_hmac/stsschema v0.0.0
+	github.com/salrashid123/aws_hmac/tpm v0.0.0
+	github.com/salrashid123/aws_hmac/tpm/signer v0.0.0
+	github.com/salrashid123/aws_hmac/tpm/signer/v4 v4.0.0
 )
 
-require (
-	github.com/davecgh/go-spew v1.1.1 // indirect
-	github.com/google/go-sev-guest v0.6.1 // indirect
-	github.com/google/logger v1.1.1 // indirect
-	github.com/google/uuid v1.3.0 // indirect
-	github.com/pborman/uuid v1.2.0 // indirect
-	github.com/pkg/errors v0.9.1 // indirect
-	golang.org/x/sys v0.8.0 // indirect
-	google.golang.org/protobuf v1.28.0 // indirect
-)
+require golang.org/x/sys v0.20.0 // indirect
 
 require (
-	github.com/aws/smithy-go v1.14.2 // indirect
-	github.com/jmespath/go-jmespath v0.4.0 // indirect
-	golang.org/x/crypto v0.6.0 // indirect
+	github.com/aws/aws-sdk-go-v2/credentials v1.17.16 // indirect
+	github.com/aws/aws-sdk-go-v2/feature/ec2/imds v1.16.3 // indirect
+	github.com/aws/aws-sdk-go-v2/internal/configsources v1.3.7 // indirect
+	github.com/aws/aws-sdk-go-v2/internal/endpoints/v2 v2.6.7 // indirect
+	github.com/aws/aws-sdk-go-v2/internal/ini v1.8.0 // indirect
+	github.com/aws/aws-sdk-go-v2/service/internal/accept-encoding v1.11.2 // indirect
+	github.com/aws/aws-sdk-go-v2/service/internal/presigned-url v1.11.9 // indirect
+	github.com/aws/aws-sdk-go-v2/service/sso v1.20.9 // indirect
+	github.com/aws/aws-sdk-go-v2/service/ssooidc v1.24.3 // indirect
+	github.com/aws/smithy-go v1.20.2 // indirect
+	golang.org/x/crypto v0.19.0 // indirect
 )
 
 replace (
