@@ -1,10 +1,12 @@
 module github.com/salrashid123/aws_hmac/tink/signer
 
-go 1.20
+go 1.21
 
-require github.com/google/tink/go v1.7.0
+toolchain go1.22.2
+
+require github.com/tink-crypto/tink-go/v2 v2.1.0
 
 require (
-	golang.org/x/crypto v0.0.0-20220214200702-86341886e292 // indirect
-	google.golang.org/protobuf v1.27.1 // indirect
+	golang.org/x/crypto v0.17.0 // indirect
+	google.golang.org/protobuf v1.33.0 // indirect
 )
