@@ -1,7 +1,9 @@
 module github.com/salrashid123/aws_hmac/tpm/signer
 
-go 1.20
+go 1.22
 
-require github.com/google/go-tpm v0.9.0
+toolchain go1.22.2
 
-require golang.org/x/sys v0.8.0 // indirect
+require github.com/google/go-tpm v0.9.1-0.20240514145214-58e3e47cd434
+
+require golang.org/x/sys v0.20.0 // indirect
